test: dedup extract_job_block into tests/common (CR-008/DRIFT-CR-008)
Move the verbatim-triplicated `extract_job_block` helper out of
`tests/ci_yml_windows_matrix.rs`, `tests/ci_gate_completeness.rs`, and
`tests/backfill_matrix_parity.rs` into a new `tests/common/yaml.rs` module
(re-exported from `tests/common/mod.rs`).

Each of the 3 files now declares `mod common; use common::yaml::extract_job_block;`
and deletes its local copy (~35 LOC each).  The canonical fn body is preserved
verbatim; 20 call sites are unaffected (identical two-arg signature).

No src/ change.  No runtime behavior change.  Test output is identical.
