+++ tests/common/mod.rs
@@ -1,2 +1,3 @@
 pub mod fixtures;
 pub mod mock_server;
+pub mod yaml;
